requirements.docker.txt
# API için gerekli temel paketler
fastapi==0.109.0
uvicorn==0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# PDF işleme
pdfplumber>=0.10.0

# Veri işleme
pandas>=2.0.0
numpy>=1.24.0

# ChromaDB ve embedding
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Ollama client
ollama>=0.1.0

# HTTP client
httpx>=0.25.0
